requirements.txt
flask-restful
flask-sqlalchemy
flask_httpauth
flask_cors